+++ src/index.d.ts
@@ -1,3 +1,0 @@
-export * from './methods';
-export * from './objects';
-export { Telegram } from './telegram';