+++ src/index.ts
@@ -1,0 +1,3 @@
+export { Telegram } from './telegram';
+export * from './types/objects';
+export * from './types/methods';