+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # @botcreate/telegram-sdk
 
+## 0.2.0
+
+### Minor Changes
+
+- Now with comments from Telegram Bot API
+
+## 0.1.10
+
+### Patch Changes
+
+- Types now in other folder
+
+## 0.1.9
+
+### Patch Changes
+
+- Fix
+
+## 0.1.8
+
+### Patch Changes
+
+- Fix
+
+## 0.1.7
+
+### Patch Changes
+
+- Fixed exports
+
+## 0.1.6
+
+### Patch Changes
+
+- Export for object interfaces
+
 ## 0.1.5
 
 ### Patch Changes